Separate Halo function and reuse it for comparison line as well
